added secret for git checkout

+++ bin/generate_go_client.sh
@@ -56,6 +56,11 @@ echo "-----------------------"
 echo "***********************"
 
 git clone https://${GITHUB_TOKEN}@github.com/dhontecillas/teamworkapigoclient.git $TMPDIR
+if [ "$?" -ge "0" ]; then
+    echo "cannot checkout repository"
+    exit 1
+fi
+
 cd $TMPDIR
 
 rm -rf pmv1
